requirements.txt
Flask==2.0.1
gunicorn
line-bot-sdk==1.19.0
beautifulsoup4==4.9.3
flasgger==0.9.5
copyheaders ==0.0.2
pymongo[srv] ==4.2.0
flask_cors